Cast as float for proper division

+++ calc.py
@@ -7,7 +7,7 @@ import sys
 
 if __name__ == '__main__':
 	command = sys.argv[1]
-	nums = [int(a) for a in sys.argv[2:]]
+	nums = [float(a) for a in sys.argv[2:]]
 	if (command == 'add'):
 		print sum(nums)
 	elif command == 'multiply':
